requirements.txt
streamlit
opencv-python-headless
numpy
Pillow
pytesseract
google-generativeai
python-dotenv
